refactor: remove duplicate functions
